param `auto_init` for repo and fix race conditions for branch protection
Basically when you create a new repo you can have a race condition where it wants to create the branch protection. The problem is that you have two dependencies:
- the repo itself
- the branch to be protected needs to exist

By adding an explicit depends we can ensure that only after the repo resource has been run will it attempt to setup branch protections.

Also in the case of wanting to create a repository and copy the history of another project in you can specify `auto_init` and it will not attempt to initialize the repo with a commit or setup protection. After copying the history you can flip `auto_init` to `true` which will flip the branch protection on without creating any commits.

Signed-off-by: Ben Abrams <me@benabrams.it>